Split the large application helper into smaller parts.
