+++ app/helpers/application_sidebar_helper.rb
@@ -1,0 +1,57 @@
+# Helpers for displaying sidebars in pages.
+module ApplicationSidebarHelper
+  # Checks if the current page has a sidebar.
+  #
+  # @return [boolean] True if there is a sidebar for the current page.
+  def has_sidebar?
+    content_for?(:layout_sidebar)
+  end
+
+  # Sets the current page to have a sidebar.
+  #
+  # The view must still render the sidebar in its content area.
+  def sidebar!
+    content_for(:layout_sidebar) do
+      'true'
+    end
+  end
+
+  # Defines the sidebar for the current page.
+  #
+  # @param [Array<String>] classes An array of classes to apply to the sidebar container.
+  # @return [String] The buffer containing the markup for the sidebar.
+  def sidebar(classes: ['col-xs-7', 'col-sm-3', 'col-md-2'])
+    sidebar!
+
+    render layout: 'layouts/sidebar', locals: { sidebar_classes: classes } do
+      yield
+    end
+  end
+
+  # Display the given sidebar items.
+  #
+  # @param [Array] items An array of sidebar items to be displayed.
+  # @param [Array<String>] classes An array of classes to apply to the sidebar items container.
+  # @return [String] The HTML string which will display the sidebar items.
+  def sidebar_items(items, classes: ['nav', 'nav-pills', 'nav-stacked'])
+    links = items.map { |item| link_to_sidebar_item(item) }
+    content_tag(:ul, class: classes) do
+      links.map { |link| concat(content_tag(:li, link)) }
+    end
+  end
+
+  # Generates the link to the given sidebar item.
+  #
+  # @param item The sidebar item.
+  # @return [String] The HTML string which will display the link.
+  def link_to_sidebar_item(item)
+    link_to(item[:path]) do
+      concat(item[:title])
+      concat(
+        content_tag(:span, class: ['unread']) do
+          badge(item[:unread]) if item[:unread] && item[:unread] > 0
+        end
+      )
+    end
+  end
+end
